fix(oc): output control correctly determines end of period (#765)
* Output control was not correctly determining if it was the end of a stress period
* Output control now uses tdis%endofperiod to determine if last time step of stress period
* Close #764
* Update to doxygen comments for output control modules
* Remove two unused UZF variables

+++ autotest/test_gwf_multimvr.py
@@ -56,7 +56,7 @@ x = [round(x, 3) for x in np.linspace(50.0, 45.0, ncolp)]
 topp = np.repeat(x, nrowp).reshape((15, 15)).T
 z = [round(z, 3) for z in np.linspace(50.0, 0.0, nlayp + 1)]
 botmp = [topp - z[len(z) - 2], topp - z[len(z) - 3], topp - z[0]]
-idomainp = np.ones((nlayp, nrowp, ncolp), dtype=np.int)
+idomainp = np.ones((nlayp, nrowp, ncolp), dtype=np.int32)
 # Zero out where the child grid will reside
 idomainp[0:2, 6:11, 2:8] = 0
